+++ ESOF-322/esof322.a3/src/src/esof322/a4/AbstractFactory.java
@@ -1,4 +1,4 @@
-package src.esof322.a3;
+package src.esof322.a4;
 
 public abstract class AbstractFactory {
 	abstract Adventure getAdventure(String adventureChoice);
