+++ ESOF-322/esof322.a3/src/src/esof322/a4/BoringRoom.java
@@ -1,4 +1,4 @@
-package src.esof322.a3;
+package src.esof322.a4;
 
 public class BoringRoom extends Room {
