+++ ESOF-322/esof322.a3/src/src/esof322/a4/CaveSite.java
@@ -1,8 +1,6 @@
 // Team Workiva adapted and modified this file.
 // Team Workiva is MacKenzie O'Bleness, Lisa Peters & Larry Lynn. 
-package src.esof322.a3;
-
-import java.io.Serializable;
+package src.esof322.a4;
 
 /**
  * Adventure Game Program Code Copyright (c) 1999 James M. Bieman
@@ -11,9 +9,11 @@ import java.io.Serializable;
  * 
  * The main routine is AdventureGame.main
  * 
+ * 
  **/
 
-// class Treasure
+// interface CaveSite
 
-public class Treasure extends Item implements Serializable {
+public interface CaveSite {
+	String enter(Player p);
 }
