+++ ESOF-322/esof322.a3/src/src/esof322/a4/Door.java
@@ -1,6 +1,6 @@
 // Team Workiva adapted and modified this file.
 // Team Workiva is MacKenzie O'Bleness, Lisa Peters & Larry Lynn. 
-package src.esof322.a3;
+package src.esof322.a4;
 
 import java.io.Serializable;
