+++ ESOF-322/esof322.a3/src/src/esof322/a4/HiddenItemsRoom.java
@@ -1,4 +1,4 @@
-package src.esof322.a3;
+package src.esof322.a4;
 
 import java.util.ArrayList;
 import java.util.HashMap;
