+++ ESOF-322/esof322.a3/src/src/esof322/a4/Item.java
@@ -1,6 +1,8 @@
 // Team Workiva adapted and modified this file.
 // Team Workiva is MacKenzie O'Bleness, Lisa Peters & Larry Lynn. 
-package src.esof322.a3;
+package src.esof322.a4;
+
+import java.io.Serializable;
 
 /**
  * Adventure Game Program Code Copyright (c) 1999 James M. Bieman
@@ -11,7 +13,22 @@ package src.esof322.a3;
  * 
  **/
 
-// class Key
+// class Item
 
-public class Key extends Item {
+public class Item implements Serializable {
+
+	private String description;
+
+	public void setDesc(String d) {
+		description = d;
+	}
+
+	public String getDesc() {
+		return description;
+	}
+
+	public String toString() {
+		return description;
+	}
+
 }
