+++ ESOF-322/esof322.a3/src/src/esof322/a4/StartGameView.java
@@ -1,4 +1,4 @@
-package src.esof322.a3;
+package src.esof322.a4;
 
 import java.io.IOException;
 import javax.swing.*;
